Add files via upload
A very basic Chatbot which tells you about the different hereos of Dota 2. The designed bot is a very conventional bot which is system-driven and responds via scrapping a background file provided to it.
The chatbox uses the chatbox file to provide information, using nltk pacakge in Python. Initial steps to learn NLP.